moved facebook button into success view, cleaned up code, removed console log, updated and cleaned up test
