+++ src/app/tell-us/success-view/success-view.component.html
@@ -29,6 +29,13 @@
   <dd>{{ success.form_version }}</dd>
 </dl>
 
+<button [disabled]="sdkStatus === false"
+  (click)="onSocialClick($event)"
+  class="fb-share-button">
+  <img src="../../../assets/facebook_share_button.jpg"
+    alt="Share">
+</button>
+
 <ng-template #dash>
   <dd>&ndash;</dd>
 </ng-template>
