+++ src/app/tell-us/success-view/success-view.component.scss
@@ -1,0 +1,9 @@
+.fb-share-button {
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  height: 38;
+  object-fit: contain;
+  outline: none;
+  width: 100px;
+}
