--- src/app/tell-us/tell-us/tell-us.component.scss
@@ -1,9 +1,0 @@
-.fb-share-button {
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-  height: 38;
-  object-fit: contain;
-  outline: none;
-  width: 100px;
-}
